+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Yellowbox Changelog
+## 0.7.5
+### Added
+* support for python 3.11
+* added the `fake_gcs` extra, to set up a fake google cloud storage container.
+### Internal
+* removed deprecated function from hvac tests
 ## 0.7.4
 ### Fixed
 * all dependencies have been unlocked for major changes (`^X` => `>=X`)